CostModel.ComputeAllocation: move key functions to separate file; add unmounted PVCs to unmounted PVs; address TODOs
